refactor: move iterators to iterators.go
