+++ mempool/clist_mempool_test.go
@@ -171,7 +171,7 @@ func TestReapMaxBytesMaxGas(t *testing.T) {
 
 	// Ensure gas calculation behaves as expected
 	checkTxs(t, mp, 1)
-	iter := mp.NewWRRIterator()
+	iter := NewWRRIterator(mp)
 	tx0 := iter.Next()
 	require.NotNil(t, tx0)
 	require.Equal(t, tx0.GasWanted(), int64(1), "transactions gas was set incorrectly")
